Split v2 lookup tests into multiple files
